fix: convert numpy types to native Python floats for JSON serialization
Fixed 500 Internal Server Error caused by numpy.float32 values that
couldn't be serialized to JSON by FastAPI.

Changes:
- Convert energy_kwh prediction to Python float using float()
- Convert carbon_kg emissions to Python float
- Ensures proper JSON serialization in API responses

Technical details:
- XGBoost model returns numpy.float32 types
- FastAPI's jsonable_encoder cannot serialize numpy types
- Solution: explicit conversion to native Python float

Result: Predictions now return proper JSON with numerical values

+++ src/api/services/ml_service.py
@@ -176,8 +176,8 @@ class MLService:
         # Transform back to original scale
         energy_kwh = np.expm1(log_prediction)  # Inverse of log1p
 
-        # Ensure non-negative
-        energy_kwh = max(0, energy_kwh)
+        # Ensure non-negative and convert to native Python float
+        energy_kwh = float(max(0, energy_kwh))
 
         # Metadata
         metadata = {
@@ -239,8 +239,8 @@ class MLService:
         source_factors = emission_factors.get(energy_source.lower(), emission_factors["grid"])
         emission_factor = source_factors.get(region, source_factors.get("default", 0.385))
 
-        # Calculate emissions
-        carbon_kg = energy_kwh * emission_factor
+        # Calculate emissions and convert to native Python float
+        carbon_kg = float(energy_kwh * emission_factor)
 
         return carbon_kg, emission_factor
